API code refactor
