+++ rackio/api/tags.py
@@ -1,17 +1,16 @@
 # -*- coding: utf-8 -*-
 """rackio/api/tags.py
 
-This module implements all class Resources for the Tag Engine.
+This module implements all Tag Resources for the Tag Engine.
 """
 
 import json
 
 from .core import RackioResource
-from .hook import rackio_hook
-from .auth_hook import Authorize, authorize
+from .auth_hook import authorize
 
 from ..dao import TagsDAO
-from ..managers.auth import SYSTEM_ROLE, VISITOR_ROLE
+from ..managers.auth import SYSTEM_ROLE, ADMIN_ROLE, VISITOR_ROLE
 
 
 class BaseResource(RackioResource):
@@ -21,7 +20,7 @@ class BaseResource(RackioResource):
 
 class TagCollectionResource(BaseResource):
 
-    @authorize([SYSTEM_ROLE, VISITOR_ROLE])
+    @authorize([SYSTEM_ROLE, ADMIN_ROLE, VISITOR_ROLE])
     def on_get(self, req, resp):
 
         doc = self.dao.get_all()
@@ -31,14 +30,14 @@ class TagCollectionResource(BaseResource):
 
 class TagResource(BaseResource):
 
-    @authorize([SYSTEM_ROLE, VISITOR_ROLE])
+    @authorize([SYSTEM_ROLE, ADMIN_ROLE, VISITOR_ROLE])
     def on_get(self, req, resp, tag_id):
 
         doc = self.dao.get(tag_id)
 
         resp.body = json.dumps(doc, ensure_ascii=False)
 
-    @authorize([SYSTEM_ROLE])
+    @authorize([SYSTEM_ROLE, ADMIN_ROLE])
     def on_post(self, req, resp, tag_id):
         
         value = req.media.get('value')
@@ -76,65 +75,4 @@ class TagResource(BaseResource):
             }
         
         resp.body = json.dumps(doc, ensure_ascii=False)
-
-
-class TagHistoryResource(BaseResource):
-
-    def on_get(self, req, resp, tag_id):
-
-        doc = self.dao.get_history(tag_id)
-
-        resp.body = json.dumps(doc, ensure_ascii=False)
-
-    
-class TrendResource(BaseResource):
-
-    def on_post(self, req, resp, tag_id):
-
-        tstart = req.media.get('tstart')
-        tstop = req.media.get('tstop')
-
-        doc = self.dao.get_trend(tag_id, tstart, tstop)
-
-        resp.body = json.dumps(doc, ensure_ascii=False)
-
-
-class TrendCollectionResource(BaseResource):
-
-    def on_post(self, req, resp):
-
-        tags = req.media.get('tags')
-
-        tstart = req.media.get('tstart')
-        tstop = req.media.get('tstop')
-    
-        result = self.dao.get_trends(tags, tstart, tstop)
-
-        resp.body = json.dumps(result, ensure_ascii=False)
-
-
-class WaveformResource(BaseResource):
-
-    def on_post(self, req, resp, tag_id):
-
-        tstart = req.media.get('tstart')
-        tstop = req.media.get('tstop')
-
-        doc = self.dao.get_waveform(tag_id, tstart, tstop)
-
-        resp.body = json.dumps(doc, ensure_ascii=False)
-
-
-class WaveformCollectionResource(BaseResource):
-
-    def on_post(self, req, resp):
-
-        tags = req.media.get('tags')
-
-        tstart = req.media.get('tstart')
-        tstop = req.media.get('tstop')
-    
-        result = self.dao.get_waveforms(tags, tstart, tstop)
-
-        resp.body = json.dumps(result, ensure_ascii=False)
